+++ src/lib.rs
@@ -27,6 +27,7 @@ pub mod test;
 pub mod watch;
 
 #[cfg(feature = "uuid")]
+#[doc(hidden)]
 #[deprecated(since = "0.11.1", note = "use `clickhouse::serde::uuid` instead")]
 pub mod uuid {
     pub use crate::serde::uuid::*;